// types.ts
export interface TestimonialType {
  id: number;
  name: string;
  role: string;
  image: string;
  quote: string;
  rating: number;
}

export interface ModuleType {
  id: number;
  title: string;
  description: string;
  iconName: string;
}

export interface BenefitType {
  title: string;
  description: string;
  iconName: string;
}